# Conflicts:
#	dangjia-group/dangjia-master/dangjia-service-master/src/main/java/com/dangjia/acg/service/repair/MendOrderCheckService.java

+++ dangjia-group/dangjia-goods/dangjia-api-goods/src/main/java/com/dangjia/acg/dto/budget/BudgetDTO.java
@@ -2,6 +2,7 @@ package com.dangjia.acg.dto.budget;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -13,7 +14,7 @@ import java.util.List;
 public class BudgetDTO {
     private Double workerPrice;//人工
     private Double caiPrice;//材料服务
-    private Double totalPrice;//总计
+    private BigDecimal totalPrice;//总计
 
     private List<BudgetItemDTO> budgetItemDTOList;
